Split countries/utils.js into smaller files
