+++ src/app/countries/country.js
@@ -1,0 +1,32 @@
+/*
+ * Copyright (C) 2019 The "mysteriumnetwork/mysterium-vpn" Authors.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+// @flow
+
+import { QualityLevel } from 'mysterium-vpn-js'
+
+type Country = {
+  id: string,
+  code: ?string,
+  name: string,
+  isFavorite: boolean,
+  quality: number | null,
+  qualityLevel: QualityLevel,
+  trusted: boolean
+}
+
+export type { Country }
