+++ src/app/countries/utils.js
@@ -18,21 +18,10 @@
 // @flow
 import countries from './list'
 import ProposalDTO from 'mysterium-tequilapi/lib/dto/proposal'
-import type { FavoriteProviders } from '../user-settings/user-settings'
-import { Metrics, QualityCalculator, QualityLevel } from 'mysterium-vpn-js'
+import type { Country } from './country'
 
-const COUNTRY_NAME_UNRESOLVED = 'N/A'
 const COUNTRY_CODE_LENGTH = 2
-
-type Country = {
-  id: string,
-  code: ?string,
-  name: string,
-  isFavorite: boolean,
-  quality: number | null,
-  qualityLevel: QualityLevel,
-  trusted: boolean
-}
+const COUNTRY_NAME_UNRESOLVED = 'N/A'
 
 function getCountryLabel (country: Country, maxNameLength: ?number = null, maxIdentityLength: ?number = 9) {
   const identity = limitedLengthString(country.id, maxIdentityLength)
@@ -44,13 +33,6 @@ function getCountryLabel (country: Country, maxNameLength: ?number = null, maxId
   return `${title} (${identity})`
 }
 
-function getSortedCountryListFromProposals (
-  proposals: Array<ProposalDTO>,
-  favorites: FavoriteProviders): Array<Country> {
-  const countries = proposals.map(getCountryFromProposal).map(countryFavoriteMapper(favorites))
-  return countries.sort(compareCountries)
-}
-
 function limitedLengthString (value: string, maxLength: ?number = null): string {
   if (maxLength && value.length > maxLength) {
     return value.substring(0, maxLength) + '..'
@@ -58,10 +40,9 @@ function limitedLengthString (value: string, maxLength: ?number = null): string 
   return value
 }
 
-function countryFavoriteMapper (favorites: FavoriteProviders): (Country) => Country {
-  return (country: Country) => {
-    return { ...country, isFavorite: favorites.has(country.id) }
-  }
+function isCountryKnown (countryCode: ?string): boolean {
+  return countryCode != null &&
+    typeof countries[countryCode.toLocaleLowerCase()] !== 'undefined'
 }
 
 function isProposalTrusted (proposal: ProposalDTO): boolean {
@@ -74,113 +55,9 @@ function isProposalTrusted (proposal: ProposalDTO): boolean {
   return true
 }
 
-function getCountryFromProposal (proposal: ProposalDTO): Country {
-  const calculator = new QualityCalculator()
-  const quality = calculator.calculateValue(getMetrics(proposal))
-  const qualityLevel = calculator.calculateLevel(quality)
-  const trusted = isProposalTrusted(proposal)
-
-  return {
-    id: proposal.providerId,
-    code: getCountryCodeFromProposal(proposal),
-    name: getCountryNameFromProposal(proposal),
-    isFavorite: false,
-    quality,
-    qualityLevel,
-    trusted
-  }
-}
-
-function compareCountries (a: Country, b: Country) {
-  if (a.isFavorite && !b.isFavorite) {
-    return -1
-  }
-  if (!a.isFavorite && b.isFavorite) {
-    return 1
-  }
-
-  if (a.quality !== null || b.quality !== null) {
-    if (a.quality !== null && b.quality !== null) {
-      if (a.quality < b.quality) {
-        return 1
-      }
-      if (b.quality < a.quality) {
-        return -1
-      }
-    } else {
-      if (a.quality === null) {
-        return 1
-      } else {
-        return -1
-      }
-    }
-  }
-
-  if (a.name > b.name) {
-    return 1
-  }
-  if (b.name > a.name) {
-    return -1
-  }
-
-  return 0
-}
-
-function isCountryKnown (countryCode: ?string): boolean {
-  return countryCode != null &&
-    typeof countries[countryCode.toLocaleLowerCase()] !== 'undefined'
-}
-
-function getCountryName (countryCode: string): string {
-  countryCode = countryCode.toLowerCase()
-  if (!isCountryKnown(countryCode)) {
-    return COUNTRY_NAME_UNRESOLVED
-  }
-
-  return countries[countryCode]
-}
-
-function getCountryNameFromProposal (proposal: ProposalDTO): string {
-  const countryCode = getCountryCodeFromProposal(proposal)
-  if (!countryCode) {
-    return COUNTRY_NAME_UNRESOLVED
-  }
-
-  return getCountryName(countryCode)
-}
-
-function getCountryCodeFromProposal (proposal: ProposalDTO): ?string {
-  if (proposal.serviceDefinition == null) {
-    return null
-  }
-  if (proposal.serviceDefinition.locationOriginate == null) {
-    return null
-  }
-  if (proposal.serviceDefinition.locationOriginate.country == null) {
-    return null
-  }
-
-  return proposal.serviceDefinition.locationOriginate.country
-}
-
-function getMetrics (proposal: ProposalDTO): Metrics {
-  if (!proposal.metrics || !proposal.metrics.connectCount) {
-    return { connectCount: { success: 0, fail: 0, timeout: 0 } }
-  }
-  const connectCount = proposal.metrics.connectCount
-  return {
-    connectCount: {
-      success: connectCount.success,
-      fail: connectCount.fail,
-      timeout: connectCount.timeout
-    }
-  }
-}
-
-export type { Country }
 export {
   getCountryLabel,
-  getSortedCountryListFromProposals,
   isCountryKnown,
-  isProposalTrusted
+  isProposalTrusted,
+  COUNTRY_NAME_UNRESOLVED
 }
