+++ src/app/data-fetchers/country-list.js
@@ -16,9 +16,9 @@
  */
 
 // @flow
-import type { Country } from '../countries/utils'
+import type { Country } from '../countries/country'
 import Publisher from '../../libraries/publisher'
-import { getSortedCountryListFromProposals } from '../countries/utils'
+import { getSortedCountryListFromProposals } from '../countries/parsing'
 import ProposalDTO from 'mysterium-tequilapi/lib/dto/proposal'
 import type { Subscriber } from '../../libraries/publisher'
 import type { FavoriteProviders } from '../user-settings/user-settings'
